Extract ReplicationSmoothing from ShipController into ReplicationSmoothing.cs
